+++ apps/user-front/src/app/layout.tsx
@@ -5,9 +5,8 @@ import { ReactNode, Suspense } from 'react';
 
 import { Metadata } from 'next';
 
+import { Providers } from './providers';
 import Analytics from '@/components/GA/Analytics';
-import { AuthProvider } from '@/components/Provider/AuthProvider';
-import { ReactQueryProvider } from '@/components/Provider/ReactQueryProvider';
 import { GA_TRACKING_ID } from '@/libs/ga/gtag';
 
 export const metadata: Metadata = {
@@ -41,15 +40,13 @@ export default function RootLayout({ children }: { children: ReactNode }) {
           />
         </>
       )}
-      <body className='min-h-screen'>
-        <ReactQueryProvider>
-          <Suspense fallback={<div>페이지를 불러오는 중입니다...</div>}>
-            <AuthProvider>
-              <main className='flex flex-col h-[calc(100vh-120px)]'>{children}</main>
-              <Analytics />
-            </AuthProvider>
-          </Suspense>
-        </ReactQueryProvider>
+      <body>
+        <Suspense>
+          <Providers>
+            {children}
+            <Analytics />
+          </Providers>
+        </Suspense>
       </body>
     </html>
   );
