+++ apps/user-front/src/mocks/handlers.ts
@@ -1,1 +1,14 @@
-export const handlers = [];
+import { mockNotificationListResponse } from '@repo/api/user/notifications';
+import { http, HttpResponse } from 'msw';
+
+export const handlers = [
+  http.get('/user/notifications', () => {
+    return HttpResponse.json(
+      mockNotificationListResponse({
+        page: 1,
+        size: 20,
+        sort: ['sentAt,desc'],
+      }),
+    );
+  }),
+];
