+++ apps/user-front/src/tests/setup.ts
@@ -10,6 +10,22 @@ beforeAll(() => {
   vi.mock('next/navigation', () => import('next-router-mock'));
 });
 
+beforeEach(() => {
+  const mockIntersectionObserver = vi.fn();
+
+  mockIntersectionObserver.mockReturnValue({
+    observe: vi.fn(),
+    unobserve: vi.fn(),
+    disconnect: vi.fn(),
+  });
+
+  Object.defineProperty(window, 'IntersectionObserver', {
+    writable: true,
+    configurable: true,
+    value: mockIntersectionObserver,
+  });
+});
+
 afterEach(() => server.resetHandlers());
 
 afterAll(() => server.close());
